admin_panel/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-jwt-extended==4.5.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
pytest==7.3.1
pytest-cov==4.0.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
flask-cors==4.0.0
